requirements.txt
accelerate==1.3.0
aiohttp==3.9.1
aioprometheus==23.12.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.3.post2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cupy-cuda12x==12.1.0
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.17.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
effdet==0.4.1
einops==0.7.0
emoji==2.9.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.2
fastrlock==0.8.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.6
iopath==0.1.10
ipykernel==6.29.0
ipython==8.20.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain==0.0.352
langchain-community==0.0.5
langchain-core==0.1.23
langchain-openai==0.0.6
langdetect==1.0.9
langsmith==0.0.87
layoutparser==0.3.4
lxml==4.9.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.12.0
opencv-python==4.8.1.78
orjson==3.9.14
packaging==23.2
pandas==2.1.4
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
pikepdf==8.10.1
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pypdf==3.17.3
pypdfium2==4.25.0
pysbd==0.3.4
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
quantile-python==1.1
ragas==0.1.1
rapidfuzz==3.5.2
ray==2.9.2
rdflib==7.0.0
referencing==0.33.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.18.0
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==3.4.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.21.0
tonic-validate==3.1.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.48.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
unstructured==0.11.6
unstructured-client==0.15.0
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
urllib3==2.1.0
uvicorn==0.27.1
uvloop==0.19.0
vllm==0.3.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
xformers==0.0.23.post1
xxhash==3.4.1
yarl==1.9.4
